refactor PlayerState
