+++ src/toniarts/openkeeper/game/state/GameState.java
@@ -159,8 +159,6 @@ public class GameState extends AbstractPauseAwareState implements IGameLogicUpda
                     GameState.this.stateManager.attach(worldState);
 
                     GameState.this.stateManager.attach(new SoundState(false));
-
-                    GameState.this.stateManager.attach(new SystemMessageState(app, false));
                     setProgress(0.60f);
 
                     GameState.this.stateManager.attach(new PartyState(false));
@@ -263,7 +261,6 @@ public class GameState extends AbstractPauseAwareState implements IGameLogicUpda
                 GameState.this.stateManager.getState(ActionPointState.class).setEnabled(true);
                 GameState.this.stateManager.getState(PartyState.class).setEnabled(true);
                 GameState.this.stateManager.getState(SoundState.class).setEnabled(true);
-                GameState.this.stateManager.getState(SystemMessageState.class).setEnabled(true);
 
                 // Set initialized
                 GameState.this.initialized = true;
